+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # ChangeLog
+- add DuckDB::LIBRARY_VERSION
 
 # 0.6.1
 - bump Ruby to 3.2.0 on CI